From Lacey and Ameya +++++++++++
So I’m in a participant looking at an auto-generated 12m Event. It says the window is 12/20/2011 – 7/20/2012.

According to us, the window should be open until 4/20/2012. 12m window are only 4 months long. Is the window for this even being generated incorrectly? The 4 is right below the 7 on the number pad.

__________________

I found the problem. In the auto_eventinfo_setting table the visit window for 12 month event needs to be changed from 11 to 18 (7months) to 11 to 15 (4 months).
I confirmed this with Lacey. Updating the MS SQL script should fix the problem.