From Ameya: November 23, 2011 10:52 AM
This is the fix for the Consent_type_CL3 code table.

I tried to recreate the relationship which was having subpanel issues. But when in redeployed the Sample Collection table, it gave me similar errors for other subpanels which were working fine previously. Just thought this info might be helpful.